Move rpm_is_older to utils.py
