+++ dirsrvtests/tests/suites/tls/mldsa_test.py
@@ -10,11 +10,9 @@ import logging
 import pytest
 import os
 import sys
-import itertools
-import rpm
 import socket
 import subprocess
-from lib389.utils import ds_is_older
+from lib389.utils import ds_is_older, rpm_is_older
 from lib389._constants import DN_DM, PW_DM, DEFAULT_SUFFIX
 from lib389.config import Encryption, CertmapLegacy
 from lib389.idm.user import UserAccount
@@ -29,25 +27,6 @@ if DEBUGGING:
 else:
     logging.getLogger(__name__).setLevel(logging.INFO)
 log = logging.getLogger(__name__)
-
-
-def rpm_is_older(pkg, version):
-    ts = rpm.TransactionSet()
-    mi = ts.dbMatch('name', pkg)
-    for h in mi:
-        print(f"{pkg} {h['version']} {version}")
-        for n1,n2 in itertools.zip_longest(h['version'].split('.'), version.split('.'), fillvalue=""):
-            try:
-                if int(n1) < int(n2):
-                    return True
-                if int(n1) > int(n2):
-                    return False
-            except ValueError:
-                if n1 < n2:
-                    return True
-                if n1 > n2:
-                    return False
-    return False
 
 
 script_content="""
